fix: Save relay source of posts again

+++ app/src/main/java/com/kaiwolfram/nozzle/data/eventProcessor/EventProcessor.kt
@@ -29,10 +29,8 @@ class EventProcessor(
     private val idCache = Collections.synchronizedSet(mutableSetOf<String>())
 
     override fun process(event: Event, relayUrl: String?) {
-        if (idCache.contains(event.id)) return
-
         if (event.isReaction()) {
-            processReaction(event = event, relayUrl = relayUrl)
+            processReaction(event = event)
             return
         }
         if (event.isPost()) {
@@ -50,44 +48,50 @@ class EventProcessor(
     }
 
     private fun processPost(event: Event, relayUrl: String?) {
-        if (!verify(event)) {
-            return
-        }
-        scope.launch {
-            postDao.insertIfNotPresent(
-                PostEntity(
-                    id = event.id,
-                    pubkey = event.pubkey,
-                    replyToId = event.getReplyId(),
-                    replyToRootId = event.getRootReplyId(),
-                    repostedId = event.getRepostedId(),
-                    content = event.content,
-                    createdAt = event.createdAt,
+        if (!verify(event)) return
+
+        if (!idCache.contains(event.id)) {
+            scope.launch {
+                postDao.insertIfNotPresent(
+                    PostEntity(
+                        id = event.id,
+                        pubkey = event.pubkey,
+                        replyToId = event.getReplyId(),
+                        replyToRootId = event.getRootReplyId(),
+                        repostedId = event.getRepostedId(),
+                        content = event.content,
+                        createdAt = event.createdAt,
+                    )
                 )
-            )
+            }
         }
+
         insertEventRelay(eventId = event.id, relayUrl = relayUrl)
+        idCache.add(event.id)
     }
 
-    private fun processReaction(event: Event, relayUrl: String?) {
+    private fun processReaction(event: Event) {
         if (event.content != "+") return
+        if (idCache.contains(event.id)) return
+        if (!verify(event)) return
 
-        if (!verify(event)) {
-            return
-        }
+        idCache.add(event.id)
+
         event.getReactedToId()?.let {
             scope.launch {
                 reactionDao.like(eventId = it, pubkey = event.pubkey)
             }
-            insertEventRelay(eventId = event.id, relayUrl = relayUrl)
         }
     }
 
     private fun processContactList(event: Event) {
-        if (!verify(event)) {
-            return
-        }
+        if (idCache.contains(event.id)) return
+        if (!verify(event)) return
+
+        idCache.add(event.id)
+
         scope.launch {
+            // TODO: dao.deleteIfOutdated():Boolean to save one call
             val latestTimestamp = contactDao.getLatestTimestamp(event.pubkey) ?: 0
             if (event.createdAt > latestTimestamp) {
                 contactDao.deleteList(pubkey = event.pubkey)
@@ -105,13 +109,18 @@ class EventProcessor(
     }
 
     private fun processMetadata(event: Event) {
+        if (idCache.contains(event.id)) return
         if (!verify(event)) {
             Log.d(TAG, "Metadata is invalid ${event.id}")
             return
         }
+
+        idCache.add(event.id)
+
         Log.d(TAG, "Process profile event ${event.content}")
         deserializeMetadata(event.content)?.let {
             scope.launch {
+                // TODO: Return if deleted to save insert call
                 profileDao.deleteIfOutdated(pubkey = event.pubkey, createdAt = event.createdAt)
                 profileDao.insertOrIgnore(
                     ProfileEntity(
@@ -132,8 +141,6 @@ class EventProcessor(
         val isValid = event.verify()
         if (!isValid) {
             Log.d(TAG, "Invalid event ${event.id} kind ${event.kind}")
-        } else {
-            idCache.add(event.id)
         }
         return isValid
     }
